LayoutTests/media/modern-media-controls/pip-support/ipad/pip-support-tap-expected.txt

Testing the PiPSupport behavior when clicking the PiP button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.webkitPresentationMode is "picture-in-picture"
PASS successfullyParsed is true

TEST COMPLETE
